Move maptiler-geocoding control to __future__
